+++ ontimize/components/bar-menu/menu-item/o-bar-menu-item.component.ts
@@ -72,7 +72,7 @@ export class OBarMenuItemComponent implements OnInit {
   public ngOnInit() {
     if (typeof (this.route) === 'string') {
       // no devuelve promise
-      // this.menu.getAuthGuardService().isRestricted(this.route)
+      // this.menu.getPermissionsService().isRestricted(this.route)
       //   .then(restricted => this.restricted = restricted)
       //   .catch(err => this.restricted = true);
     } else {